[Silver III] Title: 주유소, Time: 88 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -8,8 +8,8 @@ int main()
     int N;
     cin >> N;
 
-    vector<int> road(N - 1);
-    vector<int> cost(N);
+    vector<long long> road(N - 1);
+    vector<long long> cost(N);
 
     for (int i = 0; i < N - 1; i++)
     {
@@ -21,7 +21,7 @@ int main()
     }
 
     long long total = 0;
-    int min_cost = cost[0];
+    long long min_cost = cost[0];
 
     for (int i = 0; i < N - 1; i++)
     {
@@ -29,7 +29,7 @@ int main()
         {
             min_cost = cost[i];
         }
-        
+
         total += min_cost * road[i];
     }
